Corrigindo Bugs e adicionando funções

+++ Models/Perfil.php
@@ -9,18 +9,16 @@ class Perfil extends Model {
     public function getPerfil($id_perfil) {
         $array = array();
 
-        $sql = $this->pdo->prepare("SELECT posts.data_post, posts.mensagem, imagem_perfil.url, usuarios.id as id_usuario, usuarios.nome as nome_perfil 
-        FROM posts 
-        LEFT JOIN usuarios ON posts.id_usuario = usuarios.id 
-        LEFT JOIN imagem_perfil ON posts.id_usuario = imagem_perfil.id_usuario 
-        WHERE posts.id_usuario = :id_usuario");
-        $sql->bindValue(":id_usuario", $id_perfil);
+        $sql = $this->pdo->prepare("SELECT usuarios.id, usuarios.nome, imagem_perfil.url, posts.mensagem FROM usuarios LEFT JOIN posts ON posts.id_usuario = usuarios.id LEFT JOIN imagem_perfil ON usuarios.id = imagem_perfil.id_usuario WHERE usuarios.id = :id_perfil");
+        $sql->bindValue(":id_perfil", $id_perfil);
         $sql->execute();  
 
         if($sql->rowCount() > 0) {
             
             $array = $sql->fetchAll();
 
+        }else {
+            $array['id_usuario'] = $id_perfil;
         }
         return $array;
     }
@@ -50,9 +48,21 @@ class Perfil extends Model {
     
     public function mudarFoto($id_user, $nomedoarquivo) {
         
-        $sql = $this->pdo->prepare("DELETE FROM imagem_perfil WHERE id_usuario = :id_usuario");
+        $sql = $this->pdo->prepare("SELECT url FROM imagem_perfil WHERE id_usuario = :id_usuario");
         $sql->bindValue(":id_usuario", $id_user);
         $sql->execute();
+        
+        if($sql->rowCount() > 0) {
+            $url = $sql->fetch();
+            $filename ='Images/posts/'.$url['url'];
+            
+            unlink($filename);
+            
+            $sql = $this->pdo->prepare("DELETE FROM imagem_perfil WHERE id_usuario = :id_usuario");
+            $sql->bindValue(":id_usuario", $id_user);
+            $sql->execute();
+            
+        }    
         
         $sql = $this->pdo->prepare("INSERT INTO imagem_perfil SET id_usuario = :id_usuario, url = :url");
         $sql->bindValue(":id_usuario", $id_user);
